# Fix reusing float histograms