@linasm 2023-12-14
Does this fix affect the benchmark results of https://github.com/prometheus/prometheus/pull/13276?