@fpetkovski 2023-12-15
There seems to be a small difference, but the benefits from reusing objects are still there
````
 benchstat main.out new.out                                                                                   Filips-MacBook-Pro-2.local: Fri Dec 15 07:13:14 2023

name                         old time/op    new time/op    delta
NativeHistograms/sum-8          490ms ± 5%     491ms ± 2%    ~     (p=0.690 n=5+5)
NativeHistograms/sum_rate-8     763ms ± 3%     805ms ± 1%  +5.49%  (p=0.008 n=5+5)

name                         old alloc/op   new alloc/op   delta
NativeHistograms/sum-8          414MB ± 0%     415MB ± 0%  +0.15%  (p=0.032 n=5+5)
NativeHistograms/sum_rate-8     641MB ± 0%     665MB ± 0%  +3.66%  (p=0.008 n=5+5)

name                         old allocs/op  new allocs/op  delta
NativeHistograms/sum-8          5.16M ± 0%     5.16M ± 0%    ~     (p=0.056 n=5+5)
NativeHistograms/sum_rate-8     10.3M ± 0%     11.0M ± 0%  +7.23%  (p=0.008 n=5+5)
```